+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 0.13.0
 
 * Removed `TaskWrap::Inject::Defaults`. This is now implemented through `dsl`'s `:inject` option.
+* Extract invocation logic in `TaskBuilder::Task` into `Task#call_option`.
 
 # 0.12.2